Removed some unnecessary testing files
